backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
google-generativeai==0.7.2
pydantic==2.7.1
pydantic-settings==2.3.3
python-dotenv==1.0.1